Split browser Playwright flows
